updating seeds system

+++ api/test/factories.ts
@@ -131,11 +131,13 @@ export async function truncateAll() {
     await connection.query('TRUNCATE "expenditures" RESTART IDENTITY CASCADE');
     await connection.query('TRUNCATE "contributions" RESTART IDENTITY CASCADE');
     await connection.query('TRUNCATE "addresses" RESTART IDENTITY CASCADE');
-    try {
-        await connection.query('TRUNCATE "gis_boundaries" RESTART IDENTITY CASCADE');
-    } catch (error) {
-        if (process.env.NODE_ENV !== 'test') {
-            console.log('table gis_boundaries not found');
+    if (process.env.NODE_ENV !== 'production') {
+        try {
+            await connection.query('TRUNCATE "gis_boundaries" RESTART IDENTITY CASCADE');
+        } catch (error) {
+            if (process.env.NODE_ENV !== 'test') {
+                console.log('table gis_boundaries not found');
+            }
         }
     }
 }
